+++ CHANGES.rst
@@ -32,6 +32,7 @@ Improvements
 - Add keyboard shortcut (CTRL-SHIFT-A) to toggle room booking admin override (:pr:`5909`)
 - Make language list accessible (:issue:`5899`, :pr:`5903`, thanks :user:`foxbunny`)
 - Add accessible label to the main page link (:issue:`5934`, :pr:`5935`, thanks :user:`foxbunny`)
+- Add bypass block links (:issue:`5932`, :pr:`5939`, thanks :user:`foxbunny`)
 
 Bugfixes
 ^^^^^^^^